finhub_puller/puller_requirements.txt
aiohttp==3.8.1
asyncpg==0.25.0
sqlalchemy==1.4.35
alembic==1.7.7
psycopg2-binary==2.9.3
python-dotenv==0.20.0
